Python Ai Chat Bot/requirements.txt
streamlit
google-generativeai
python-dotenv
openai
requests
flask
